Reviewed-on: https://codeberg.org/socialhome/socialhome/pulls/660

+++ socialhome/tasks/apps.py
@@ -1,10 +1,14 @@
+import logging
 import sys
 
 import django_rq
 from django.apps import AppConfig
+from rq.exceptions import InvalidJobOperation
+from rq.job import JobStatus
 
 from socialhome.streams.tasks import streams_tasks
 
+logger = logging.getLogger("socialhome")
 
 class TasksConfig(AppConfig):
     name = "socialhome.tasks"
@@ -19,7 +23,10 @@ class TasksConfig(AppConfig):
 
         # Delete any existing jobs in the scheduler when the app starts up
         for job in scheduler.get_jobs():
-            job.delete()
+            try:
+                if job.get_status() != JobStatus.SCHEDULED: job.delete()
+            except InvalidJobOperation as ex:
+               logger.warning("TasksConfig.ready - failed to get scheduled job status: %s", ex)
 
         # Queue tasks
         streams_tasks(scheduler)
